# docs/openapi/openapi.json
{
  "paths": {
...
    "/api/v1/students/": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "sort_by",
            "schema": {
              "type": "string",
              "default": "created_at"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "description": "Page number (1-indexed)",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "required": false
          },
          {
            "in": "query",
            "name": "per_page",
            "description": "Results per page, max 100",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "required": false
          },
          {
            "in": "query",
            "name": "department",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "division",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "search",
            "description": "Search by name or roll number",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "min_attendance",
            "description": "Filter by min attendance %",
            "schema": {
              "type": "number"
            },
            "required": false
          }
        ],
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "List students",
        "description": "Returns a paginated list of students. Faculty can only see their own division. Admin sees all.",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "responses": {
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "409": {
            "description": "Roll number already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Create student",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/students/{student_id}": {
      "get": {
        "responses": {
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Get student by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "student_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ],
      "patch": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdate"
              }
            }
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Partially update student",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error1"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Soft-delete student",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/students/{student_id}/attendance": {
      "get": {
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSummary"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Get student attendance summary",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "student_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
    "/api/v1/students/{student_id}/results": {
      "get": {
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResultResponse"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Students"
        ],
        "summary": "Get student results",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "student_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
...
    "/api/v1/attendance/student/{student_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSummary"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "tags": [
          "Attendance"
        ],
        "summary": "Get student attendance summary and history",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "student_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
...
  },
  "info": {
    "title": "DY Patil University ERP API",
    "version": "v1"
  },
...
  "openapi": "3.1.0",
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Error code"
          },
          "status": {
            "type": "string",
            "description": "Error name"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "object",
            "description": "Errors"
          }
        }
      },
...
      "Error1": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Not found"
          },
          "code": {
            "type": "string",
            "example": "STUDENT_NOT_FOUND"
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "default": null
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          },
          "has_next": {
            "type": "boolean"
          },
          "has_prev": {
            "type": "boolean"
          }
        }
      },
      "StudentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "roll": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "division": {
            "type": "string"
          },
          "semester": {
            "type": "integer"
          },
          "phone": {
            "type": "string"
          },
          "attendance_pct": {
            "type": "number",
            "default": 0.0,
            "description": "Current month attendance %"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StudentCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "roll": {
            "type": "string",
            "example": "CS2024001"
          },
          "department": {
            "type": "string",
            "enum": [
              "CS",
              "IT",
              "ENTC",
              "MECH",
              "CIVIL",
              "AIDS",
              "AIML"
            ]
          },
          "year": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4
          },
          "division": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D"
            ]
          },
          "semester": {
            "type": "integer",
            "minimum": 1,
            "maximum": 8
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?[0-9]{10,13}$"
          },
          "parent_phone": {
            "type": "string",
            "pattern": "^\\+?[0-9]{10,13}$"
          },
          "address": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "department",
          "division",
          "dob",
          "email",
          "name",
          "roll",
          "semester",
          "year"
        ]
      },
      "StudentUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "AttendanceSummary": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "integer"
          },
          "student_name": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "attended": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "percentage": {
            "type": "number"
          }
        }
      },
      "ResultResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "student_id": {
            "type": "integer"
          },
          "semester": {
            "type": "string"
          },
          "subject_id": {
            "type": "integer"
          },
          "internal_marks": {
            "type": "number"
          },
          "external_marks": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "grade": {
            "type": "string"
          },
          "is_published": {
            "type": "boolean"
          }
        }
      },
...
    },
    "responses": {
      "DEFAULT_ERROR": {
        "description": "Default error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UNPROCESSABLE_ENTITY": {
        "description": "Unprocessable Entity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}